+++ src/symbols_linux.cpp
@@ -481,7 +481,7 @@ void ElfParser::parseDebugFrameSection() {
             DwarfParser dwarf(_cc->name(), _base, section_start, section_start + debug_frame_section->sh_size);
             _cc->setDwarfTable(dwarf.table(), dwarf.count());
             Log::warn("*****************For file %s", _cc->name());
-            for (int i = 0; i < dwarf.count(); i++) {
+            for (u64 i = 0; i < dwarf.count(); i++) {
                 Log::warn("Dwarf table idx = %d, loc = %u, cfa = %d, fp_off = %d, pc_off = %d", i, dwarf.table()[i].loc,
                 dwarf.table()[i].cfa, dwarf.table()[i].fp_off, dwarf.table()[i].pc_off);
             }
